deleted initial files and add splash screen
