+++ app/index.jsx
@@ -1,0 +1,31 @@
+import { useEffect } from 'react'
+import { View, StyleSheet } from 'react-native'
+import { useRouter } from 'expo-router'
+import HueLogo from '../components/HueLogo'
+
+export default function SplashScreen() {
+  const router = useRouter()
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      router.replace('/day')
+    }, 2000)
+
+    return () => clearTimeout(timer)
+  }, [])
+
+  return (
+    <View style={styles.container}>
+      <HueLogo />
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+})
